Set test case to uint8

+++ test/test_image.py
@@ -81,7 +81,7 @@ class ImageTester(unittest.TestCase):
             jpeg_bytes = encode_jpeg(img, quality=75)
             with open(expected_file, 'rb') as f:
                 pil_bytes = f.read()
-                pil_bytes = torch.as_tensor(list(pil_bytes))
+                pil_bytes = torch.as_tensor(list(pil_bytes), dtype=torch.uint8)
             self.assertTrue(jpeg_bytes.equal(pil_bytes))
 
         with self.assertRaisesRegex(ValueError, "Expected a torch.uint8 tensor."):
